actually, set_meta_class helper is needed on boot - move it under Util
